refactor choices
